+++ src/index.html
@@ -8,6 +8,6 @@
     </style>
     <body>
         <div id="app"></div>
-        <script type="module" src="./index.ts"></script>
+        <script type="module" src="./index.tsx"></script>
     </body>
 </html>
