+++ src/index.tsx
@@ -3,4 +3,4 @@ import ReactDOM from 'react-dom'
 
 import App from './App'
 
-ReactDOM.render( <App/>, document.getElementById('app'),);
+ReactDOM.render(<App />, document.getElementById('app'))
